refactor: move DeleteStmt to variables.py and change List to Iterable
