+++ src/astx/callables.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import Any, Iterable, List, Optional, cast
+from typing import Any, Iterable, Optional, cast
 
 from public import public
 
@@ -12,7 +12,6 @@ from astx.base import (
     ASTNodes,
     DataType,
     Expr,
-    Identifier,
     ReprStruct,
     SourceLocation,
     StatementType,
@@ -415,39 +414,6 @@ class YieldFromExpr(Expr):
         """Return the AST structure of the object."""
         key = "YIELDFROM-EXPR"
         value = self.value.get_struct(simplified)
-        return self._prepare_struct(key, value, simplified)
-
-
-class DeleteStmt(StatementType):
-    """AST class for 'del' statements."""
-
-    value: List[Identifier]
-
-    def __init__(
-        self,
-        value: List[Identifier],
-        loc: SourceLocation = NO_SOURCE_LOCATION,
-        parent: Optional[ASTNodes] = None,
-    ) -> None:
-        """Initialize the DeleteStmt instance."""
-        super().__init__(loc=loc, parent=parent)
-        self.value = value
-        self.kind = ASTKind.DeleteStmtKind
-
-    def __str__(self) -> str:
-        """Return a string representation of the object."""
-        value_str = ", ".join(str(value) for value in self.value)
-        return f"DeleteStmt[{value_str}]"
-
-    def get_struct(self, simplified: bool = False) -> ReprStruct:
-        """Return the AST structure of the object."""
-        key = "DELETE"
-        # Convert list to a dictionary or other ReprStruct-compatible format
-        value: ReprStruct = {
-            f"target_{i}": val.get_struct(simplified)
-            for i, val in enumerate(self.value)
-        }
-
         return self._prepare_struct(key, value, simplified)
