+++ tests/test_callables.py
@@ -2,13 +2,11 @@
 
 import pytest
 
-from astx.base import Identifier
 from astx.blocks import Block
 from astx.callables import (
     Argument,
     Arguments,
     AwaitExpr,
-    DeleteStmt,
     FunctionAsyncDef,
     FunctionCall,
     FunctionDef,
@@ -176,25 +174,3 @@ def test_yieldfrom_expr() -> None:
     assert yieldfrom_expr.get_struct()
     assert yieldfrom_expr.get_struct(simplified=True)
     visualize(yieldfrom_expr.get_struct())
-
-
-def test_delete_stmt() -> None:
-    """Test `DeleteStmt` class."""
-    var1 = Identifier(value="x")
-    var2 = Identifier(value="y")
-
-    # Create a DeleteStmt with multiple values
-    delete_stmt = DeleteStmt(value=[var1, var2])
-
-    # Test basic properties
-    assert str(delete_stmt)
-    assert delete_stmt.get_struct()
-    assert delete_stmt.get_struct(simplified=True)
-
-    # Optional: Visualize for documentation
-    visualize(delete_stmt.get_struct())
-
-    # Test with a single value
-    single_delete = DeleteStmt(value=[var1])
-    assert str(single_delete)
-    assert single_delete.get_struct()
